RAG_Chunk_Code/requirements.txt
# RAG System Dependencies

# Core
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# LLM Providers
groq>=0.4.0
requests>=2.28.0

# Optional: Gemini
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
